+++ CHANGELOG.md
@@ -10,6 +10,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
  - Add desktop integration for Windows executables, which can now be launched using Protontricks
  - Add `protontricks-desktop-install` to install desktop integration for the local user. This is only necessary if the installation method doesn't do this automatically.
 
+### Changed
+ - Improve GUI dialog. The prompt to select the Steam app now uses zenity's "list" dialog, which supports scrolling and search.
+
 ### Fixed
  - Display proper error messages in certain cases when corrupted VDF files are found